requirements.txt
absl-py==2.0.0
accelerate==0.20.3
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
autocorrect==2.6.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.10
bypy==1.6.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
deprecate==1.0
dill==0.3.7
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.16.4
hydra-core==1.3.2
HyperPyYAML==1.2.1
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.0.1
inexactsearch==1.0.2
ipykernel==6.25.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.0.post2
lightgbm==4.0.0
lightning-utilities==0.9.0
llvmlite==0.40.1
loguru==0.7.2
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
polars==0.19.8
pooch==1.6.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.0.9
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.7.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
rapidfuzz==3.3.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soundex==1.1.3
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
speechbrain==0.5.15
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
textgen==1.1.1
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchmetrics==0.7.0
torchvision==0.10.1+cu111
tornado==6.3.2
tqdm==4.66.0
traitlets==5.9.0
transformers==4.33.3
triton==2.1.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.0
widgetsnbextension==4.0.8
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
